+++ piecrust/tests/constructor.rs
@@ -20,36 +20,36 @@ fn constructor() -> Result<(), Error> {
         ModuleData::builder(OWNER).constructor_arg(&0xabu8),
     )?;
 
-    assert_eq!(session.query::<(), u8>(id, "read_value", &())?, 0xab);
+    assert_eq!(session.call::<(), u8>(id, "read_value", &())?, 0xab);
 
     // perform transaction and make sure that the contract works as expected
-    session.transact::<(), ()>(id, "increment", &())?;
-    assert_eq!(session.query::<(), u8>(id, "read_value", &())?, 0xac);
+    session.call::<(), ()>(id, "increment", &())?;
+    assert_eq!(session.call::<(), u8>(id, "read_value", &())?, 0xac);
 
     // we should not be able to call init directly
-    let result = session.transact::<u8, ()>(id, CONTRACT_INIT_METHOD, &0xaa);
+    let result = session.call::<u8, ()>(id, CONTRACT_INIT_METHOD, &0xaa);
     assert!(
         result.is_err(),
         "calling init directly as transaction should not be allowed"
     );
     // we should not be able to call init as query neither
-    let result = session.query::<u8, ()>(id, CONTRACT_INIT_METHOD, &0xaa);
+    let result = session.call::<u8, ()>(id, CONTRACT_INIT_METHOD, &0xaa);
     assert!(
         result.is_err(),
         "calling init directly as query should not be allowed"
     );
 
     // make sure the state is still ok
-    assert_eq!(session.query::<(), u8>(id, "read_value", &())?, 0xac);
+    assert_eq!(session.call::<(), u8>(id, "read_value", &())?, 0xac);
 
     // initialized state should live through across session boundaries
     let commit_id = session.commit()?;
     let mut session = vm.session(SessionData::builder().base(commit_id))?;
-    assert_eq!(session.query::<(), u8>(id, "read_value", &())?, 0xac);
+    assert_eq!(session.call::<(), u8>(id, "read_value", &())?, 0xac);
 
     // not being able to call init directly should also be enforced across
     // session boundaries
-    let result = session.transact::<u8, ()>(id, CONTRACT_INIT_METHOD, &0xae);
+    let result = session.call::<u8, ()>(id, CONTRACT_INIT_METHOD, &0xae);
     assert!(
         result.is_err(),
         "calling init directly should never be allowed"